Min and max property
Added min and max property for all pickers, number picker has number
values and date/time pickers have string values.
