+++ NumericSpinerAdvance/doc/Spinner.html
@@ -44,7 +44,7 @@
     <h2>Constructor</h2>
     
 
-    <h4 class="name" id="Spinner"><span class="type-signature"></span>new Spinner<span class="signature">(inputElement, buttonElement, width, spinnerType, scrollable, precision, step, value, dateFormat, stepUnit)</span><span class="type-signature"></span></h4>
+    <h4 class="name" id="Spinner"><span class="type-signature"></span>new Spinner<span class="signature">(inputElement, buttonElement, width, spinnerType, scrollable, precision, step, value, dateFormat, stepUnit, min, max)</span><span class="type-signature"></span></h4>
 
     
 
@@ -314,6 +314,52 @@
         </tr>
 
     
+
+        <tr>
+            
+                <td class="name"><code>min</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">any</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">Minimal value</td>
+        </tr>
+
+    
+
+        <tr>
+            
+                <td class="name"><code>max</code></td>
+            
+
+            <td class="type">
+            
+                
+<span class="param-type">any</span>
+
+
+            
+            </td>
+
+            
+
+            
+
+            <td class="description last">Maximal value</td>
+        </tr>
+
+    
     </tbody>
 </table>
 
@@ -357,7 +403,7 @@
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="numeric.spinner.advance.js.html">numeric.spinner.advance.js</a>, <a href="numeric.spinner.advance.js.html#line326">line 326</a>
+        <a href="numeric.spinner.advance.js.html">numeric.spinner.advance.js</a>, <a href="numeric.spinner.advance.js.html#line394">line 394</a>
     </li></ul></dd>
     
 
@@ -413,13 +459,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="NumericButton.html">NumericButton</a></li><li><a href="Spinner.html">Spinner</a></li><li><a href="SpinnerEvents.html">SpinnerEvents</a></li></ul><h3>Global</h3><ul><li><a href="global.html#__extends">__extends</a></li><li><a href="global.html#assembleMomentDate">assembleMomentDate</a></li><li><a href="global.html#changeValue">changeValue</a></li><li><a href="global.html#mousedown">mousedown</a></li><li><a href="global.html#mouseup">mouseup</a></li><li><a href="global.html#mousewheel">mousewheel</a></li><li><a href="global.html#Numeric">Numeric</a></li><li><a href="global.html">render</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="NumericButton.html">NumericButton</a></li><li><a href="Spinner.html">Spinner</a></li><li><a href="SpinnerEvents.html">SpinnerEvents</a></li></ul><h3>Global</h3><ul><li><a href="global.html#__extends">__extends</a></li><li><a href="global.html#assembleMomentDate">assembleMomentDate</a></li><li><a href="global.html#changeValue">changeValue</a></li><li><a href="global.html#checkRange">checkRange</a></li><li><a href="global.html#mousedown">mousedown</a></li><li><a href="global.html#mouseup">mouseup</a></li><li><a href="global.html#mousewheel">mousewheel</a></li><li><a href="global.html#Numeric">Numeric</a></li><li><a href="global.html">render</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-alpha13</a> on Mon Jan 26 2015 19:21:07 GMT+0100 (Central Europe Standard Time)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-alpha13</a> on Tue Jan 27 2015 22:56:37 GMT+0100 (Central Europe Standard Time)
 </footer>
 
 <script> prettyPrint(); </script>
