+++ NumericSpinerAdvance/doc/index.html
@@ -50,13 +50,13 @@
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="NumericButton.html">NumericButton</a></li><li><a href="Spinner.html">Spinner</a></li><li><a href="SpinnerEvents.html">SpinnerEvents</a></li></ul><h3>Global</h3><ul><li><a href="global.html#__extends">__extends</a></li><li><a href="global.html#assembleMomentDate">assembleMomentDate</a></li><li><a href="global.html#changeValue">changeValue</a></li><li><a href="global.html#mousedown">mousedown</a></li><li><a href="global.html#mouseup">mouseup</a></li><li><a href="global.html#mousewheel">mousewheel</a></li><li><a href="global.html#Numeric">Numeric</a></li><li><a href="global.html">render</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Classes</h3><ul><li><a href="NumericButton.html">NumericButton</a></li><li><a href="Spinner.html">Spinner</a></li><li><a href="SpinnerEvents.html">SpinnerEvents</a></li></ul><h3>Global</h3><ul><li><a href="global.html#__extends">__extends</a></li><li><a href="global.html#assembleMomentDate">assembleMomentDate</a></li><li><a href="global.html#changeValue">changeValue</a></li><li><a href="global.html#checkRange">checkRange</a></li><li><a href="global.html#mousedown">mousedown</a></li><li><a href="global.html#mouseup">mouseup</a></li><li><a href="global.html#mousewheel">mousewheel</a></li><li><a href="global.html#Numeric">Numeric</a></li><li><a href="global.html">render</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-alpha13</a> on Mon Jan 26 2015 19:21:07 GMT+0100 (Central Europe Standard Time)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.3.0-alpha13</a> on Tue Jan 27 2015 22:56:37 GMT+0100 (Central Europe Standard Time)
 </footer>
 
 <script> prettyPrint(); </script>
