+++ NumericSpinerAdvance/numeric.spiner.css
@@ -1,0 +1,36 @@
+input.spiner {
+    width: 90px;
+    text-align: right;
+    display: inline-block;
+}
+.input-group-btn-vertical {
+    position: relative;
+    white-space: nowrap;
+    width: 1%;
+    vertical-align: middle;
+    display: inline-block;
+}
+.input-group-btn-vertical > .btn {
+    display: block;
+    float: none;
+    width: 100%;
+    max-width: 100%;
+    padding: 8px;
+    margin-left: -1px;
+    position: relative;
+    border-radius: 0;
+}
+.input-group-btn-vertical > .btn:first-child {
+    margin-top: -4px;
+    height: 18px;
+    border-top-right-radius: 4px;
+}
+.input-group-btn-vertical > .btn:last-child {
+    height: 18px;
+    border-bottom-right-radius: 4px;
+}
+.input-group-btn-vertical i{
+    position: absolute;
+    top: 0;
+    left: 4px;
+}
