+++ NumericSpinerAdvance/src/typing/jquery.d.ts
